074.-Politicas-de-acceso-a-roles

+++ app/Http/Controllers/Admin/RolesController.php
@@ -17,6 +17,8 @@ class RolesController extends Controller
      */
     public function index()
     {
+        $this->authorize('view', new Role);
+
         return view('admin.roles.index', [
             'roles' => Role::all()
         ]);
@@ -29,8 +31,10 @@ class RolesController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', $role = new Role);
+
         return view('admin.roles.create', [
-            'role' => new Role(),
+            'role' => $role,
             'permissions' => Permission::pluck('name', 'id'),
         ]);
     }
@@ -43,6 +47,8 @@ class RolesController extends Controller
      */
     public function store(SaveRolesRequest $request)
     {
+        $this->authorize('create', new Role);
+
         $role = Role::create($request->validated());
 
         if ($request->has('permissions')) {
@@ -53,17 +59,6 @@ class RolesController extends Controller
     }
 
     /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
      * Show the form for editing the specified resource.
      *
      * @param  int  $id
@@ -71,6 +66,8 @@ class RolesController extends Controller
      */
     public function edit(Role $role)
     {
+        $this->authorize('update', $role);
+
         return view('admin.roles.edit', [
             'role' => $role,
             'permissions' => Permission::pluck('name', 'id'),
@@ -86,6 +83,8 @@ class RolesController extends Controller
      */
     public function update(SaveRolesRequest $request, Role $role)
     {
+        $this->authorize('update', $role);
+
         $role->update($request->validated());
 
         $role->permissions()->detach();
@@ -105,11 +104,7 @@ class RolesController extends Controller
      */
     public function destroy(Role $role)
     {
-        if ($role->id === 1) {
-            throw new \Illuminate\Auth\Access\AuthorizationException(__('This role can not be deleted'));
-        }
-
-        // $this->authorize('delete', $role);
+        $this->authorize('delete', $role);
 
         $role->delete();
